不能放到带空格的目录

+++ upigcmd.rb
@@ -69,7 +69,7 @@ EOF
 
 html_header = html_header.to_utf8
 html_footer = html_footer.to_utf8
-
+keyword.strip!
 File.open(options[:temp], 'w'){|temp_file|
   File.open(keyword, 'r'){|f|
     src_str = f.read
@@ -85,6 +85,7 @@ File.open(options[:temp], 'w'){|temp_file|
     src_str_utf8.each_line {|line|
       line.gsub!('　'.to_utf8, '  '.to_utf8)
       line.lstrip!
+      #todo escape html code
       temp_file.print '<p>'+line+'</p>'
     }
 
@@ -94,7 +95,16 @@ File.open(options[:temp], 'w'){|temp_file|
   }
 }
 
+script_path = File.expand_path(File.dirname(__FILE__))
+if File.exist?('xiang.css') && !File.exist?('upig2prc.exe')
+  $stderr.puts 'txt目录下不能有xiang.css'
+  exit 
+end
+File.copy(File.join(script_path, 'xiang.css'), 'xiang.css')
 result = `temp/kindlegen.exe #{options[:temp]}`
 $stderr.puts result if result.include?('Error')
 File.delete(options[:temp])
+if File.exist?('xiang.css') && !File.exist?('upig2prc.exe')
+  File.delete('xiang.css')
+end
